Delete the 8 dormant mobile task-detail skins; relocate MobileStickyBar
ADR-0058 is Accepted, so pages/taskDetail/mobileArchetypes/* is no longer a
revert path. The folder's shared.tsx was the one piece that could not go with
it: MobileStickyBar is imported by eight LIVE mobile faction pages, which are
now its only consumers, so it moves into
pages/factionDetail/mobileArchetypes/shared.tsx. Its sibling export
MobileStickyCaption had no consumer outside the deleted skins and goes with
them.
