+++ frontend/src/pages/factionDetail/mobileArchetypes/EphemeristsFactionPage.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import PraxisCard from '../../../components/PraxisCard'
 import { factionName, factionDescription } from '../../../utils/factions'
 import { EphemeristsSigil } from '../../../components/cards/ephemeristsAtoms'
-import { MobileStickyBar } from '../../taskDetail/mobileArchetypes/shared'
+import { MobileStickyBar } from './shared'
 import type { CharacterOut } from '../../../api/auth'
 import type { FactionDetailState } from '../useFactionDetail'
