--- frontend/src/pages/taskDetail/mobileArchetypes/CovenTaskDetail.tsx
@@ -1,344 +1,0 @@
-import type { CSSProperties, ReactNode } from 'react'
-import { Link } from 'react-router-dom'
-import { useTranslation } from 'react-i18next'
-import PraxisCard from '../../../components/PraxisCard'
-import { MobileStickyBar, MobileStickyCaption } from './shared'
-import { LevelJumpBanner } from '../archetypes/shared'
-import type { TaskDetailState } from '../useTaskDetail'
-
-/**
- * Cozy Coven MOBILE task-detail skin (#531) — the quest.exe window on a
- * phone. A pink scrapbook window hero (traffic-light title bar, dotted board,
- * inner notepad) carries the Caveat title + sparks/level plates, then the "what
- * we're asking" brief, the love-so-far aggregate, the spells cast (completions),
- * and a sparkle-stamped **sticky action bar** ("join the party") pinned above the
- * tab bar. Single-column, no fixed-px grid. Reuses the same {@link TaskDetailState},
- * the shared mobile sticky bar, and the `wow.*` copy + `--faction-coven-*` tokens as
- * the desktop archetype and the other COVEN pilots. Always-light: COVEN is native-light.
- */
-
-const PINK = 'var(--faction-coven)'
-const PINK_DEEP = 'var(--faction-coven-card-accent)'
-const TITLE_TEXT = 'var(--faction-coven-title-text)'
-const CARD_TEXT = 'var(--faction-coven-card-text)'
-const CARD_MUTED = 'var(--faction-coven-card-muted)'
-const WIN_BORDER = 'var(--faction-coven-win-border)'
-const NOTEPAD_BG = 'var(--faction-coven-notepad-bg)'
-const NOTEPAD_BORDER = 'var(--faction-coven-notepad-border)'
-const BODY_BG = 'var(--faction-coven-body-bg)'
-const DOT = 'var(--faction-coven-dot)'
-const SCRIPT = 'var(--faction-coven-card-font)' // Caveat
-const BODY = 'var(--font-body)' // Courier Prime
-const ON_ACCENT = 'var(--color-text-on-accent)'
-
-/** The kit's four-point sparkle. */
-function Sparkle({ size, color, style }: { size: number; color: string; style?: CSSProperties }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" style={style} aria-hidden>
-      <path
-        d="M12 0c.9 7 4.1 10.2 11 11-6.9.8-10.1 4-11 11-.9-7-4.1-10.2-11-11C7.9 10.2 11.1 7 12 0Z"
-        fill={color}
-      />
-    </svg>
-  )
-}
-
-/** Caveat section heading with a sparkle and a dashed running rule. */
-function SectionHead({ title, trailing }: { title: string; trailing?: ReactNode }) {
-  return (
-    <div style={{ display: 'flex', alignItems: 'center', gap: "var(--space-md)", marginBottom: "var(--space-md)" }}>
-      <span className="content-title" style={{ display: 'inline-flex', alignItems: 'center', gap: "var(--space-sm)", fontFamily: SCRIPT, color: TITLE_TEXT, whiteSpace: 'nowrap' }}>
-        <Sparkle size={13} color={PINK} />
-        {title}
-      </span>
-      <span style={{ flex: 1, height: 2, background: `repeating-linear-gradient(90deg, ${PINK} 0 8px, transparent 8px 14px)` }} />
-      {trailing}
-    </div>
-  )
-}
-
-export default function CovenTaskDetail({ state }: { state: TaskDetailState }) {
-  const { t } = useTranslation('tasks')
-  const {
-    task,
-    submissions,
-    mySubmission,
-    isInProgress,
-    inProgressPraxisId,
-    canSignUp,
-    slotsOpen,
-    maxTaskSlots,
-    modifiedPoints,
-    topScore,
-    voteCount,
-    sortedSubmissions,
-    submissionSort,
-    setSubmissionSort,
-    signupError,
-    handleSignup,
-    handleDrop,
-  } = state
-
-  if (!task) return null
-
-  const topId = submissions.length
-    ? submissions.reduce((a, b) => ((b.score ?? 0) > (a.score ?? 0) ? b : a)).id
-    : null
-
-  return (
-    <div
-      data-skin="coven"
-      className="py-4"
-      style={{
-        fontFamily: BODY,
-        color: CARD_TEXT,
-        background: BODY_BG,
-        backgroundImage: `radial-gradient(${DOT} 1.4px, transparent 1.4px)`,
-        backgroundSize: '15px 15px',
-        marginLeft: -16,
-        marginRight: -16,
-        paddingLeft: "var(--space-lg)",
-        paddingRight: "var(--space-lg)",
-      }}
-    >
-      {/* Breadcrumb */}
-      <nav className="mb-3" style={{ fontSize: "var(--text-sm)", letterSpacing: '0.12em', textTransform: 'uppercase', color: CARD_MUTED }}>
-        <Link to="/tasks" style={{ color: PINK_DEEP, textDecoration: 'none' }}>
-          {t('coven.breadcrumb')}
-        </Link>
-        <span aria-hidden style={{ margin: "0 var(--space-sm)", color: PINK }}>›</span>
-        <span style={{ fontFamily: SCRIPT, fontSize: "var(--text-xl)", color: TITLE_TEXT }}>{task.title}</span>
-      </nav>
-
-      {/* Hero — quest.exe window */}
-      <section
-        style={{
-          borderRadius: 16,
-          overflow: 'hidden',
-          border: `2px solid ${WIN_BORDER}`,
-          boxShadow: `0 8px 22px color-mix(in srgb, ${PINK} 22%, transparent)`,
-          marginBottom: "var(--space-xl)",
-        }}
-      >
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            gap: "var(--space-sm)",
-            padding: "var(--space-sm) var(--space-md)",
-            background: 'linear-gradient(180deg, var(--faction-coven-title-from), var(--faction-coven-title-to))',
-            borderBottom: `2px solid ${WIN_BORDER}`,
-          }}
-        >
-          {['var(--faction-coven-scrap-deep)', 'var(--faction-coven-tape)', 'var(--faction-coven-ivy-leaf)'].map((c) => (
-            <span key={c} style={{ width: 9, height: 9, borderRadius: '50%', background: c, border: '1.2px solid rgba(255,255,255,0.7)' }} />
-          ))}
-          <span style={{ marginLeft: "auto", display: 'inline-flex', alignItems: 'center', gap: "var(--space-xs)", fontFamily: SCRIPT, fontSize: "var(--text-content)", color: TITLE_TEXT }}>
-            <Sparkle size={11} color={TITLE_TEXT} />
-            {t('coven.window')}
-          </span>
-        </div>
-        <div
-          style={{
-            position: 'relative',
-            padding: "var(--space-md)",
-            background: BODY_BG,
-            backgroundImage: `radial-gradient(${DOT} 1.4px, transparent 1.4px)`,
-            backgroundSize: '13px 13px',
-          }}
-        >
-          <div style={{ background: NOTEPAD_BG, border: `1.5px solid ${NOTEPAD_BORDER}`, borderRadius: 9, padding: "var(--space-lg) var(--space-lg) var(--space-xl)" }}>
-            <div style={{ fontSize: "var(--text-sm)", letterSpacing: '0.16em', textTransform: 'uppercase', color: CARD_MUTED, marginBottom: "var(--space-sm)" }}>
-              {task.status === 'active' ? t('coven.statusOpen') : task.status}
-            </div>
-            <h1 className="content-title" style={{ fontFamily: SCRIPT, lineHeight: 0.95, color: TITLE_TEXT, margin: "0 0 var(--space-lg)", overflowWrap: 'anywhere' }}>
-              {task.title}
-            </h1>
-            <div className="flex items-center gap-2 flex-wrap">
-              <span style={{ display: 'inline-flex', alignItems: 'center', gap: "var(--space-sm)", fontSize: "var(--text-base)", letterSpacing: '0.06em', textTransform: 'uppercase', color: CARD_TEXT, background: BODY_BG, border: `1.5px solid ${NOTEPAD_BORDER}`, borderRadius: 999, padding: "var(--space-sm) var(--space-md)" }}>
-                <Sparkle size={10} color={PINK} />
-                {t('coven.level', { level: task.level_required })}
-              </span>
-              <span className="content-title" style={{ display: 'inline-flex', alignItems: 'center', gap: "var(--space-sm)", fontFamily: SCRIPT, color: PINK }}>
-                {t('coven.sparks', { points: modifiedPoints })}
-              </span>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      {/* What we're asking — brief */}
-      <section style={{ marginBottom: "var(--space-xl)" }}>
-        <SectionHead title={t('coven.askingHeading')} />
-        <div style={{ background: NOTEPAD_BG, border: `1.5px solid ${NOTEPAD_BORDER}`, borderRadius: 9, padding: "var(--space-lg) var(--space-lg)" }}>
-          <p className="content-text" style={{ fontFamily: BODY, lineHeight: 1.75, color: task.description ? CARD_TEXT : CARD_MUTED, margin: 0, whiteSpace: 'pre-wrap' }}>
-            {task.description || t('coven.askingEmpty')}
-          </p>
-        </div>
-      </section>
-
-      {/* The love so far — read-only aggregate */}
-      <section style={{ marginBottom: "var(--space-xl)" }}>
-        <SectionHead title={t('coven.loveHeading')} />
-        {voteCount > 0 ? (
-          <div style={{ display: 'flex', alignItems: 'flex-end', gap: "var(--space-md)" }}>
-            <span className="content-title" style={{ fontFamily: SCRIPT, lineHeight: 0.8, color: PINK }}>{topScore}</span>
-            <div style={{ paddingBottom: "var(--space-xs)" }}>
-              <div className="content-text" style={{ fontFamily: SCRIPT, color: TITLE_TEXT }}>{t('coven.love.top')}</div>
-              <div style={{ fontSize: "var(--text-base)", color: CARD_MUTED, marginTop: "var(--space-xs)" }}>{t('coven.love.adored', { count: voteCount })}</div>
-            </div>
-          </div>
-        ) : (
-          <p className="content-text" style={{ fontFamily: SCRIPT, color: PINK, margin: 0 }}>{t('coven.love.none')}</p>
-        )}
-      </section>
-
-      {/* Spells cast (completions) */}
-      <section>
-        <SectionHead
-          title={t('coven.spellsHeading', { count: submissions.length })}
-          trailing={
-            <div style={{ display: 'flex' }}>
-              {(['score', 'recent'] as const).map((sort) => {
-                const on = submissionSort === sort
-                return (
-                  <button
-                    key={sort}
-                    onClick={() => setSubmissionSort(sort)}
-                    style={{
-                      fontFamily: BODY,
-                      fontSize: "var(--text-xs)",
-                      letterSpacing: '0.1em',
-                      textTransform: 'uppercase',
-                      padding: "var(--space-xs) var(--space-md)",
-                      borderRadius: on ? 8 : 0,
-                      background: on ? `linear-gradient(180deg, ${PINK}, ${PINK_DEEP})` : 'transparent',
-                      color: on ? ON_ACCENT : CARD_MUTED,
-                      border: `1.5px solid ${on ? PINK_DEEP : NOTEPAD_BORDER}`,
-                      cursor: 'pointer',
-                    }}
-                  >
-                    {sort === 'score' ? t('coven.sort.mostLoved') : t('coven.sort.recent')}
-                  </button>
-                )
-              })}
-            </div>
-          }
-        />
-        {sortedSubmissions.length === 0 ? (
-          <p className="content-text" style={{ fontFamily: SCRIPT, color: PINK, margin: 0 }}>{t('coven.empty')}</p>
-        ) : (
-          <>
-            <div className="flex flex-col gap-4">
-              {sortedSubmissions.slice(0, 4).map((s) => (
-                <div key={s.id} style={{ position: 'relative', paddingTop: s.id === topId ? 'var(--space-lg)' : 0 }}>
-                  {s.id === topId && (
-                    <div
-                      style={{
-                        position: 'absolute',
-                        top: -4,
-                        left: '50%',
-                        transform: 'translateX(-50%) rotate(-2deg)',
-                        zIndex: 3,
-                        display: 'inline-flex',
-                        alignItems: 'center',
-                        gap: "var(--space-xs)",
-                        whiteSpace: 'nowrap',
-                        background: PINK,
-                        color: ON_ACCENT,
-                        fontFamily: SCRIPT,
-                        fontSize: "var(--text-xl)",
-                        padding: "var(--space-xs) var(--space-lg)",
-                        borderRadius: 14,
-                        boxShadow: '2px 3px 0 var(--faction-coven-scrap-deep)',
-                      }}
-                    >
-                      <span style={{ fontSize: "var(--text-lg)", lineHeight: 1 }}>⚜</span>
-                      {t('coven.mostLoved')}
-                    </div>
-                  )}
-                  <PraxisCard praxis={s} />
-                </div>
-              ))}
-            </div>
-            {submissions.length > 4 && (
-              <div style={{ marginTop: "var(--space-lg)" }}>
-                <Link to={`/praxes?task_id=${task.id}`} className="content-text" style={{ fontFamily: SCRIPT, color: PINK, textDecoration: 'none' }}>
-                  {t('coven.viewAll', { count: submissions.length })}
-                </Link>
-              </div>
-            )}
-          </>
-        )}
-      </section>
-
-      {/* Sticky action bar */}
-      {canSignUp && (
-        <MobileStickyBar>
-          <LevelJumpBanner state={state} />
-          {signupError && (
-            <div style={{ fontFamily: BODY, fontSize: "var(--text-lg)", color: 'var(--color-danger)', padding: "var(--space-sm) var(--space-md)", background: 'var(--faction-coven-light)', border: `1.5px solid ${NOTEPAD_BORDER}`, borderRadius: 9 }}>
-              {signupError}
-            </div>
-          )}
-          <button
-            onClick={handleSignup}
-            style={{
-              width: '100%',
-              display: 'inline-flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              gap: "var(--space-sm)",
-              background: `linear-gradient(180deg, ${PINK}, ${PINK_DEEP})`,
-              color: ON_ACCENT,
-              fontFamily: BODY,
-              fontSize: "var(--text-lg)",
-              fontWeight: 700,
-              letterSpacing: '0.1em',
-              textTransform: 'uppercase',
-              padding: "var(--space-lg) var(--space-xl)",
-              border: `1.5px solid ${PINK_DEEP}`,
-              borderRadius: 14,
-              cursor: 'pointer',
-            }}
-          >
-            <Sparkle size={13} color={ON_ACCENT} />
-            {t('coven.signup.cta', { points: modifiedPoints })}
-          </button>
-          <MobileStickyCaption color={CARD_MUTED}>
-            {t('coven.signup.slots', { open: slotsOpen, max: maxTaskSlots })}
-          </MobileStickyCaption>
-        </MobileStickyBar>
-      )}
-
-      {mySubmission && (
-        <MobileStickyBar style={{ flexDirection: 'row', alignItems: 'center' }}>
-          <span className="content-text" style={{ fontFamily: SCRIPT, color: PINK, flex: 1 }}>{t('coven.submitted.text')}</span>
-          <Link
-            to={`/praxes/${mySubmission.id}/edit`}
-            style={{ fontFamily: BODY, fontSize: "var(--text-base)", letterSpacing: '0.12em', textTransform: 'uppercase', padding: "var(--space-md) var(--space-lg)", color: ON_ACCENT, background: `linear-gradient(180deg, ${PINK}, ${PINK_DEEP})`, border: `1.5px solid ${PINK_DEEP}`, borderRadius: 12, textDecoration: 'none' }}
-          >
-            {t('coven.submitted.edit')}
-          </Link>
-        </MobileStickyBar>
-      )}
-
-      {!mySubmission && isInProgress && inProgressPraxisId !== null && (
-        <MobileStickyBar style={{ flexDirection: 'row', alignItems: 'center', gap: "var(--space-md)" }}>
-          <button
-            onClick={handleDrop}
-            style={{ background: 'none', border: 'none', cursor: 'pointer', fontFamily: SCRIPT, fontSize: "var(--text-content)", color: CARD_MUTED }}
-          >
-            {t('coven.inProgress.drop')}
-          </button>
-          <Link
-            to={`/praxes/${inProgressPraxisId}/edit`}
-            style={{ marginLeft: "auto", fontFamily: BODY, fontSize: "var(--text-md)", letterSpacing: '0.12em', textTransform: 'uppercase', padding: "var(--space-md) var(--space-xl)", color: ON_ACCENT, background: `linear-gradient(180deg, ${PINK}, ${PINK_DEEP})`, border: `1.5px solid ${PINK_DEEP}`, borderRadius: 14, textDecoration: 'none' }}
-          >
-            {t('coven.inProgress.continue')}
-          </Link>
-        </MobileStickyBar>
-      )}
-    </div>
-  )
-}
